Stub out variation functionality on windows/macos
