+++ components/canvas/canvas_data.rs
@@ -62,7 +62,6 @@ impl UnshapedTextRun<'_> {
             word_spacing,
             script: self.script,
             flags: ShapingFlags::empty(),
-            variation_settings: Default::default(),
         };
         let glyphs = font.shape_text(self.string, &options);
         Some(TextRun { font, glyphs })
