+++ components/fonts/platform/freetype/font.rs
@@ -80,7 +80,7 @@ impl PlatformFontMethods for PlatformFont {
         let data: &[u8] = font_data.as_ref();
         let face = FreeTypeFace::new_from_memory(&library, data)?;
 
-        let normalized_variations = unsafe { set_variations_for_font(face, variations, &library)? };
+        let normalized_variations = face.set_variations_for_font(variations, &library)?;
 
         let (requested_face_size, actual_face_size) = match requested_size {
             Some(requested_size) => (requested_size, face.set_size(requested_size)?),
@@ -106,7 +106,7 @@ impl PlatformFontMethods for PlatformFont {
 
         let face = FreeTypeFace::new_from_file(&library, &filename, font_identifier.index())?;
 
-        let normalized_variations = unsafe { set_variations_for_font(face, variations, &library)? };
+        let normalized_variations = face.set_variations_for_font(variations, &library)?;
 
         let (requested_face_size, actual_face_size) = match requested_size {
             Some(requested_size) => (requested_size, face.set_size(requested_size)?),
@@ -420,74 +420,4 @@ impl std::fmt::Debug for FreeTypeFaceTableProviderData {
     fn fmt(&self, _: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         Ok(())
     }
-}
-
-/// Applies to provided variations to the font face.
-///
-/// Returns the normalized font variations, which are clamped
-/// to fit within the range of their respective axis. Variation
-/// values for nonexistent axes are not included.
-///
-/// ## SAFETY:
-/// * `face` must point to a valid `FT_FaceRec` structure.
-unsafe fn set_variations_for_font(
-    face: FT_Face,
-    variations: &[FontVariation],
-    library: &FreeTypeLibraryHandle,
-) -> Result<Vec<FontVariation>, &'static str> {
-    if !FT_HAS_MULTIPLE_MASTERS(face) ||
-        variations.is_empty() ||
-        !servo_config::pref!(layout_variable_fonts_enabled)
-    {
-        // Nothing to do
-        return Ok(vec![]);
-    }
-
-    // Query variation axis of font
-    let mut mm_var: *mut FT_MM_Var = ptr::null_mut();
-    let result = unsafe { FT_Get_MM_Var(face, &mut mm_var as *mut _) };
-    if !result.succeeded() {
-        return Err("Failed to query font variations");
-    }
-
-    // Prepare values for each axis. These are either the provided values (if any) or the default
-    // ones for the axis.
-    let num_axis = unsafe { (*mm_var).num_axis } as usize;
-    let mut normalized_axis_values = Vec::with_capacity(variations.len());
-    let mut coords = vec![0; num_axis];
-    for (index, coord) in coords.iter_mut().enumerate() {
-        let axis_data = unsafe { &*(*mm_var).axis.add(index) };
-        let Some(variation) = variations
-            .iter()
-            .find(|variation| variation.tag == axis_data.tag as u32)
-        else {
-            *coord = axis_data.def;
-            continue;
-        };
-
-        // Freetype uses a 16.16 fixed point format for variation values
-        let shift_factor = 16.0_f32.exp2();
-        let min_value = axis_data.minimum as f32 / shift_factor;
-        let max_value = axis_data.maximum as f32 / shift_factor;
-        normalized_axis_values.push(FontVariation {
-            tag: variation.tag,
-            value: variation.value.min(max_value).max(min_value),
-        });
-
-        *coord = (variation.value * shift_factor) as i64;
-    }
-
-    // Free the MM_Var structure
-    unsafe {
-        FT_Done_MM_Var(library.freetype_library, mm_var);
-    }
-
-    // Set the values for each variation axis
-    let result =
-        unsafe { FT_Set_Var_Design_Coordinates(face, coords.len() as u32, coords.as_ptr()) };
-    if !result.succeeded() {
-        return Err("Could not set variations for font face");
-    }
-
-    Ok(normalized_axis_values)
 }
